Navigation with C-v / M-v , and buffer local font change.
The navigation keys are changed to emacs standard.

Font change is now local to the buffer. It now affects the size only.
(For me this is what I need, as I mainly show code, that has the right
highlighting already.)

Bonus: I addes some ^L to the file, so it can be used for demo purposes.